adjust canal json unit test.
